Move the codegen module into core.
